comment require AuthenticationPrincipal

+++ src/main/java/com/ouaskanas/educonnect/Web/CommentController.java
@@ -1,8 +1,11 @@
 package com.ouaskanas.educonnect.Web;
 
 import com.ouaskanas.educonnect.Dao.Entities.Comment;
+import com.ouaskanas.educonnect.Dao.Repositories.CommentRepository;
+import com.ouaskanas.educonnect.Dao.Repositories.PostRepository;
 import com.ouaskanas.educonnect.Service.Manager.CommentManager;
-import com.ouaskanas.educonnect.Service.Manager.PostManager;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -11,8 +14,14 @@ import java.util.List;
 @RequestMapping("api/v1/comment")
 public class CommentController {
 
-    private PostManager postManager;
+    private final PostRepository postRepository;
+    private final CommentRepository commentRepository;
     private CommentManager commentManager;
+
+    public CommentController(PostRepository postRepository, CommentRepository commentRepository) {
+        this.postRepository = postRepository;
+        this.commentRepository = commentRepository;
+    }
 
     @GetMapping("/{post_id}/allcomments")
     public List<Comment> getAllComments(@PathVariable long post_id) {
@@ -20,13 +29,20 @@ public class CommentController {
     }
 
     @PostMapping("/{post_id}/addcomment")
-    public Comment addComment(@PathVariable long post_id, @RequestBody Comment comment) {
-        return commentManager.postComment(post_id, comment);
+    public ResponseEntity<Comment> addComment(@PathVariable long post_id, @RequestBody Comment comment) {
+        if(!postRepository.existsById(post_id)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
+        commentManager.postComment(post_id, comment);
+        return ResponseEntity.status(HttpStatus.CREATED).body(comment);
     }
 
     @DeleteMapping("/{post_id}/{comment_id}")
-    public String deleteComment(@PathVariable long post_id, @PathVariable long comment_id) {
+    public ResponseEntity<String> deleteComment(@PathVariable long post_id, @PathVariable long comment_id) {
+        if(!postRepository.existsById(post_id) && !commentRepository.existsById(comment_id)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
         commentManager.deleteComment(post_id, comment_id);
-        return "Comment deleted";
+        return ResponseEntity.status(HttpStatus.OK).body("Comment Deleted");
     }
 }
